More xaphan helper compatibility (#4)
* Add deathSound

Adds a string for a custom audio event that plays when an enemy is killed

* Adds deathSound field

deathSound is a string for an audio event when the enemy is killed

* Update CustomEnemy to make death sounds work

tested and working!

* NPC shootSound, NPC jump deterministic randomness

The randomness is now calculated with a more TAS-friendly seed and enemies can optionally make a sound when shooting.
The comf logger has been commented out to prevent logspam in the console.

* More Xaphan Helper compatibility

Enemies can be damaged if they are attacked with a bomb/megabomb/drone beam from XaphanHelper if a hidden attribute is true. (Made specifically for celesterraria, normal users won't be able to use this feature)

* make enemy dash more lenient

---------

Co-authored-by: EllaTAS <86128357+ella-TAS@users.noreply.github.com>